Implement ScreenCaptureThumbnail component and integrate it into discussion history
